LUI-3 - removing Thing + separate the component + Test
