+++ example/index.tsx
@@ -1,7 +1,7 @@
 import 'react-app-polyfill/ie11';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Button } from '../dist';
+import { Button } from '../.';
 
 const App = () => {
   const types = [
