+++ test/blah.test.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Default as Thing } from '../stories/Thing.stories';
+import { Default as Button } from '../stories/Button.stories';
 
-describe('Thing', () => {
+describe('Button', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<Thing />, div);
+    ReactDOM.render(<Button />, div);
     ReactDOM.unmountComponentAtNode(div);
   });
 });
